+++ sorting-searching/qsort/main.c
@@ -25,10 +25,10 @@ static int intcmp(const void *a, const void *b);
 int main(void)
 {
     int arr[] = {5, 2, 8, 1, 3};
-    int size = sizeof(arr) / sizeof(arr[0]);
+    size_t size = sizeof(arr) / sizeof(arr[0]);
 
     printf("Original array: ");
-    for(int i = 0; i < size; i++)
+    for(size_t i = 0; i < size; i++)
     {
         printf("%d ", arr[i]);
     }
@@ -40,7 +40,7 @@ int main(void)
 
     printf("Sorted array: ");
 
-    for(int i = 0; i < size; i++)
+    for(size_t i = 0; i < size; i++)
     {
         printf("%d ", arr[i]);
     }
